release: v0.7.1 — kv.increment positional amount + accurate RateLimitError.retry_after
- ctx.kv.increment(key, amount) now accepts amount positionally (was keyword-only,
  inconsistent with decrement / Redis INCRBY). A third-party economy plugin crashed
  calling it positionally. Backward-compatible: keyword amount=/path= unchanged.
- RateLimitError.retry_after surfaced accurately from structured host error metadata
  (now a float); MockContext mirrors the new increment signature.

+++ yourbot_sdk/_exceptions.py
@@ -44,7 +44,7 @@ class RateLimitError(SdkError):
     """
     code = "RATE_LIMITED"
 
-    def __init__(self, message: str = "Rate limit exceeded", retry_after: int = 0,
+    def __init__(self, message: str = "Rate limit exceeded", retry_after: float = 0,
                  *, code: Optional[str] = None):
         self.retry_after = retry_after
         super().__init__(message, code=code)
